requirements.txt
aldjemy==0.8.0
backports-abc==0.5
certifi==2017.7.27.1
chardet==3.0.4
contextlib2==0.5.5
Django==1.11.4
django-crontab==0.7.1
django-mysql==2.1.0
idna==2.6
Jinja2==2.9.6
jsonfield==2.0.2
livereload==2.5.1
MarkupSafe==1.0
middleware==1.2.3
mysqlclient==1.3.10
psycopg2==2.7.3
PyExcelerate==0.7.1
python-dateutil==2.6.1
pytz==2017.2
raven==6.2.1
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.13
tornado==4.5.1
urllib3==1.22
